projects/aql-builder/src/app/shared/date-to-today.pipe.ts
/* Copyright 2021 Better Ltd (www.better.care)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Pipe, PipeTransform } from '@angular/core';

@Pipe({
  name: 'dateToToday'
})
export class DateToTodayPipe implements PipeTransform {

  transform(value: string): string {
    const today = new Date().setHours(0, 0, 0, 0);
    const date = new Date(value).setHours(0, 0, 0, 0);
    return date === today ? 'today' : value;
  }
}
